+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### enhancements
   * Improve support to Rails 4.x associations with more duck typing instead of Array checks.
   * Support Rails 4.1 and Ruby 2.1.1.
+  * Add `skip_blanks` configuration option to skip generating blank attributes
+    instead of generating them with a default message. (by github.com/moktin)
 
 ### bugfix
   * Do not yield default blank value to the block when using an empty association. (by github.com/aptinio)